+++ pymahjong/agari_checker.py
@@ -1,11 +1,11 @@
 from typing import Callable
 
-from src.deficiency_calculator import (
+from pymahjong.deficiency_calculator import (
     calculate_normal_deficiency,
     calculate_seven_pairs_deficiency,
     calculate_thirteen_orphans_deficiency,
 )
-from src.schema.count import HandCount
+from pymahjong.schema.count import HandCount
 
 
 def validate_hand_count_for_agari_checking(func: Callable[[HandCount], bool]):
