+++ pymahjong/enum/common.py
