+++ pymahjong/exception.py
