+++ pymahjong/schema/efficiency_data.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from pydantic import BaseModel
 
-from src.schema.tile import Tile
+from pymahjong.schema.tile import Tile
 
 
 class EfficiencyData(BaseModel):
