+++ pymahjong/schema/hand.py
@@ -4,9 +4,9 @@ from typing import Iterable, Optional
 
 from pydantic import BaseModel
 
-from src.enum.common import CallTypeEnum
-from src.schema.call import Call
-from src.schema.tile import Tile
+from pymahjong.enum.common import CallTypeEnum
+from pymahjong.schema.call import Call
+from pymahjong.schema.tile import Tile
 
 
 class Hand(BaseModel):
