+++ tests/schema/test_tile.py
@@ -1,7 +1,7 @@
 import pytest
 
-from src.enum.common import TileTypeEnum
-from src.schema.tile import Tile
+from pymahjong.enum.common import TileTypeEnum
+from pymahjong.schema.tile import Tile
 
 
 @pytest.mark.parametrize(
